[Refactoring ⚙️] send Comment 별도 관리
