+++ src/components/PostDetail/components/PostDetailInfo/components/PostDetailComment/components/PostCommentInput/PostCommentInput.tsx
@@ -7,19 +7,17 @@ import { POST_DETAIL_MODAL_MESSAGE } from "@/constants/modalMessage"
 import { Post } from "@/types"
 import useAuthUserStore from "@/stores/useAuthUserStore"
 import { useNavigate } from "react-router-dom"
-import { useMutation, useQueryClient } from "@tanstack/react-query"
-import { sendComment } from "@/components/PostDetail/apis/sendComment"
 import commentValidation from "./util/commentValidation"
 import { POST_DETAIL_ERROR_MESSAGE } from "@/constants/errorMessage"
 import AlertModal from "@/components/Modal/components/AlertModal/AlertModal"
-
-const SEND_MUTATION_QUERY_KEY = "SEND_MUTATION_MUTATION_KEY_9128178621782"
+import useSendComment from "@/components/PostDetail/hooks/useSendComment"
 
 interface PostCommentInputProps {
   post: Post
 }
 
 const PostCommentInput = ({ post }: PostCommentInputProps) => {
+  const { sendCommentMutate } = useSendComment()
   const [writeComment, setWriteComment] = useState("")
   const [alertMessage, setAlertMessage] = useState("")
   const { user: authUser, isLoggedIn } = useAuthUserStore()
@@ -34,20 +32,6 @@ const PostCommentInput = ({ post }: PostCommentInputProps) => {
     showModal: showAlert,
     closeModal: closeAlert,
   } = useModal()
-
-  const queryClient = useQueryClient()
-  const CommentApi_send = useMutation({
-    mutationKey: [SEND_MUTATION_QUERY_KEY],
-    mutationFn: sendComment,
-    onSuccess: () => {
-      queryClient.refetchQueries()
-      setWriteComment("")
-    },
-    onError: () => {
-      setAlertMessage(POST_DETAIL_ERROR_MESSAGE.COMMENT_SUBMIT.ERROR)
-      showAlert()
-    },
-  })
 
   const navigate = useNavigate()
 
@@ -80,10 +64,21 @@ const PostCommentInput = ({ post }: PostCommentInputProps) => {
       return
     }
 
-    CommentApi_send.mutate({
-      comment: writeComment,
-      postId: post._id,
-    })
+    sendCommentMutate.mutate(
+      {
+        comment: writeComment,
+        postId: post._id,
+      },
+      {
+        onSuccess: () => {
+          setWriteComment("")
+        },
+        onError: () => {
+          setAlertMessage(POST_DETAIL_ERROR_MESSAGE.COMMENT_SUBMIT.ERROR)
+          showAlert()
+        },
+      },
+    )
   }
 
   const handleChangeComment = ({ target }: ChangeEvent<HTMLInputElement>) => {
@@ -117,17 +112,21 @@ const PostCommentInput = ({ post }: PostCommentInputProps) => {
         </S.PostCommentInputContainer>
       </S.PostCommentInputLayout>
 
-      <ConfirmModal
-        isShow={isShowConfirm}
-        onClose={handleCloseConfirmModal}
-        message={POST_DETAIL_MODAL_MESSAGE.CONFIRM.COMMENT_FOCUS}
-      />
+      {isShowConfirm && (
+        <ConfirmModal
+          isShow={isShowConfirm}
+          onClose={handleCloseConfirmModal}
+          message={POST_DETAIL_MODAL_MESSAGE.CONFIRM.COMMENT_FOCUS}
+        />
+      )}
 
-      <AlertModal
-        isShow={isShowAlert}
-        onClose={closeAlert}
-        alertMessage={alertMessage}
-      />
+      {isShowAlert && (
+        <AlertModal
+          isShow={isShowAlert}
+          onClose={closeAlert}
+          alertMessage={alertMessage}
+        />
+      )}
     </>
   )
 }
